P10 Finalized

+++ Java Advanced/src/L03SetsAndMapsAdvancedExcercise/P10PopulationCounter.java
@@ -8,8 +8,8 @@ import java.util.Scanner;
 public class P10PopulationCounter {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
-        Map<String, Map<String, Integer>> dataTable = new LinkedHashMap<>();
-        Map<String, Integer> countries = new LinkedHashMap<>();
+        Map<String, Map<String, Long>> dataTable = new LinkedHashMap<>();
+        Map<String, Long> countries = new LinkedHashMap<>();
 
         String input = scan.nextLine();
 
@@ -17,11 +17,11 @@ public class P10PopulationCounter {
             String[] tokens = input.split("\\|");
             String city = tokens[0];
             String country = tokens[1];
-            int population = Integer.parseInt(tokens[2]);
+            Long population = Long.parseLong(tokens[2]);
 
             dataTable.putIfAbsent(country, new LinkedHashMap<>());
             dataTable.get(country).putIfAbsent(city, population);
-            countries.putIfAbsent(country, 0);
+            countries.putIfAbsent(country, 0L);
             countries.put(country, countries.get(country) + population);
 
             input = scan.nextLine();
@@ -30,16 +30,16 @@ public class P10PopulationCounter {
         printData(countries, dataTable);
     }
 
-    private static void printData(Map<String, Integer> countries, Map<String, Map<String, Integer>> dataTable) {
+    private static void printData(Map<String, Long> countries, Map<String, Map<String, Long>> dataTable) {
         countries.entrySet()
                 .stream()
-                .sorted(Map.Entry.<String,Integer>comparingByValue().reversed())
+                .sorted(Map.Entry.<String,Long>comparingByValue().reversed())
                 .forEach(country -> {
                     System.out.printf("%s (total population: %d)%n", country.getKey(), country.getValue());
-                    Map<String, Integer> cities = dataTable.get(country.getKey());
+                    Map<String, Long> cities = dataTable.get(country.getKey());
                     cities.entrySet()
                             .stream()
-                            .sorted(Map.Entry.<String,Integer>comparingByValue().reversed())
+                            .sorted(Map.Entry.<String,Long>comparingByValue().reversed())
                             .forEach(ct-> System.out.printf("=>%s: %d%n", ct.getKey(), ct.getValue()));
                 });
     }
